requirements.txt
python-dotenv==1.0.1
langchain==0.2.16
langchain-experimental==0.0.65
langchain-ibm==0.1.5
colorama==0.4.6